jadwal

+++ app/Http/Controllers/Frontend/JadwalPatroliDetailController.php
@@ -8,6 +8,7 @@ use App\Models\JadwalPatroliDetail;
 use App\Models\Lokasi;
 use App\Traits\CommonTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Yajra\DataTables\DataTables;
 
 class JadwalPatroliDetailController extends Controller
@@ -16,7 +17,7 @@ class JadwalPatroliDetailController extends Controller
      * Display a listing of the resource.
      */
 
-    use CommonTrait;   
+    use CommonTrait;
     public function table(Request $request)
     {
         if ($request->ajax()) {
@@ -25,30 +26,34 @@ class JadwalPatroliDetailController extends Controller
             $data = JadwalPatroliDetail::where('comid', $comid)->where('patroli_id', $id);
             return DataTables::of($data)
                 ->addIndexColumn()
-                ->addColumn('patroli_id', function($row){
+                ->addColumn('patroli_id', function ($row) {
                     return $row->jadwal->name ?? '';
                 })
-                ->addColumn('location_id', function($row){
+                ->addColumn('location_id', function ($row) {
                     return $row->location->nama_lokasi ?? '';
                 })
-                ->addColumn('jam_patroli', function($row){
-                    return $row->jam_awal .' - '.$row->jam_akhir;
+                ->addColumn('jam_patroli', function ($row) {
+                    return $row->jam_awal . ' - ' . $row->jam_akhir;
                 })
                 ->addColumn('action', function ($row) {
-                    
                     $button = '';
                     $button .= '<center>';
-                    $button .= '<button onclick="editData(' . $row->id . ')" title="Edit Data" class="me-0 btn btn-insoft btn-warning"><i class="bi bi-pencil-square"></i></button>';
-                    $button .= '<button onclick="deleteData(' . $row->id . ')" title="Hapus Data" class="btn btn-insoft btn-danger"><i class="bi bi-trash3"></i></button>';
+                    if (Auth::user()->level == 'owner') {
+                        $button .= '<button onclick="editData(' . $row->id . ')" title="Edit Data" class="me-0 btn btn-insoft btn-warning"><i class="bi bi-pencil-square"></i></button>';
+                        $button .= '<button onclick="deleteData(' . $row->id . ')" title="Hapus Data" class="btn btn-insoft btn-danger"><i class="bi bi-trash3"></i></button>';
+                    } else {
+                        $button .= '<button disabled title="Edit Data" class="me-0 btn btn-insoft btn-warning"><i class="bi bi-pencil-square"></i></button>';
+                        $button .= '<button disabled title="Hapus Data" class="btn btn-insoft btn-danger"><i class="bi bi-trash3"></i></button>';
+                    }
 
                     $button .= '</center>';
                     return $button;
                 })
-                
+
                 ->addColumn('comid', function ($row) {
                     return $row->company->company_name ?? '';
                 })
-                
+
                 ->rawColumns(['action'])
                 ->make(true);
 
@@ -56,12 +61,7 @@ class JadwalPatroliDetailController extends Controller
         }
     }
 
-
-
-    public function index()
-    {
-       
-    }
+    public function index() {}
 
     /**
      * Show the form for creating a new resource.
@@ -78,18 +78,23 @@ class JadwalPatroliDetailController extends Controller
     {
         $input = $request->all();
         $validated = $request->validate([
-            "patroli_id" => "required",
-            "location_id" => "required",
-            "urutan" => "required",
-            "jam_awal" => "required",
-            "jam_akhir" => "required"
-        ]);   
+            'patroli_id' => 'required',
+            'location_id' => 'required',
+            'urutan' => 'nullable',
+            'jam_awal' => 'required',
+            'jam_akhir' => 'required',
+        ]);
+
+        $lastUrutan = JadwalPatroliDetail::where('comid', $this->comid())->where('patroli_id', $request->patroli_id)->max('urutan');
+
+        $urutan = $lastUrutan ? $lastUrutan + 1 : 1;
 
         $input['comid'] = $this->comid();
+        $input['urutan'] = $request->urutan ?? $urutan;
         JadwalPatroliDetail::create($input);
         return response()->json([
-            "success" => true,
-            "message" => "Success"
+            'success' => true,
+            'message' => 'Success',
         ]);
     }
 
@@ -99,11 +104,17 @@ class JadwalPatroliDetailController extends Controller
     public function show(string $id)
     {
         $view = 'jadwal-patroli-detail';
-        $jadwal = JadwalPatroli::find($id);
-        $locations = Lokasi::where('comid', $this->comid())
-            ->where('is_active', 1)
-            ->get();
-        return view('frontend.setting.jadwal_patroli.detail.detail',compact('view','id', 'jadwal','locations'));
+        $jadwal = JadwalPatroli::findorFail($id);
+        if (!$jadwal) {
+            return redirect()->back();
+        }
+
+        if ($jadwal->comid != $this->comid()) {
+            return redirect()->back();
+        }
+
+        $locations = Lokasi::where('comid', $this->comid())->where('is_active', 1)->get();
+        return view('frontend.setting.jadwal_patroli.detail.detail', compact('view', 'id', 'jadwal', 'locations'));
     }
 
     /**
@@ -111,7 +122,8 @@ class JadwalPatroliDetailController extends Controller
      */
     public function edit(string $id)
     {
-        //
+        $data = JadwalPatroliDetail::find($id);
+        return $data;
     }
 
     /**
@@ -119,7 +131,22 @@ class JadwalPatroliDetailController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
+        $input = $request->all();
+        $validated = $request->validate([
+            'patroli_id' => 'required',
+            'location_id' => 'required',
+            'urutan' => 'required',
+            'jam_awal' => 'required',
+            'jam_akhir' => 'required',
+        ]);
+
+        $input['comid'] = $this->comid();
+        $data = JadwalPatroliDetail::find($id);
+        $data->update($input);
+        return response()->json([
+            'success' => true,
+            'message' => 'Success',
+        ]);
     }
 
     /**
@@ -127,6 +154,10 @@ class JadwalPatroliDetailController extends Controller
      */
     public function destroy(string $id)
     {
-        //
+        JadwalPatroliDetail::destroy($id);
+        return response()->json([
+            'success' => true,
+            'message' => 'sucess',
+        ]);
     }
 }
